Pass global template variables via context processor.
Avoid passing template globals to each view manually, which turns out to
be easy to forget.
