+++ regulations/templates/regulations/dap_analytics.html
@@ -2,6 +2,6 @@
 DAP Analytics tag.
 {% endcomment %}
 
-{% if ANALYTICS.DAP.DAP_URL_PARAMS %}
-  <script src="https://dap.digitalgov.gov/Universal-Federated-Analytics-Min.js?{{ ANALYTICS.DAP.DAP_URL_PARAMS }}" id="_fed_an_ua_tag"></script>
+{% if globals.ANALYTICS.DAP.DAP_URL_PARAMS %}
+  <script src="https://dap.digitalgov.gov/Universal-Federated-Analytics-Min.js?{{ globals.ANALYTICS.DAP.DAP_URL_PARAMS }}" id="_fed_an_ua_tag"></script>
 {% endif %}
