+++ regulations/templates/regulations/google_analytics.html
@@ -2,25 +2,25 @@
 Google Analytics tags. Override for different IDs
 {% endcomment %}
 
-{% if ANALYTICS.GOOGLE.GTM_SITE_ID %}
+{% if globals.ANALYTICS.GOOGLE.GTM_SITE_ID %}
   <!-- Google Tag Manager -->
-  <noscript><iframe src="https://www.googletagmanager.com/ns.html?id={{ ANALYTICS.GOOGLE.GTM_SITE_ID }}"
+  <noscript><iframe src="https://www.googletagmanager.com/ns.html?id={{ globals.ANALYTICS.GOOGLE.GTM_SITE_ID }}"
   height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
   <script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
   new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
   j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
   'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-  })(window,document,'script','dataLayer','{{ ANALYTICS.GOOGLE.GTM_SITE_ID }}');</script>
+  })(window,document,'script','dataLayer','{{ globals.ANALYTICS.GOOGLE.GTM_SITE_ID }}');</script>
   <!-- End Google Tag Manager -->
 {% endif %}
-{% if ANALYTICS.GOOGLE.GA_SITE_ID %}
+{% if globals.ANALYTICS.GOOGLE.GA_SITE_ID %}
   <script>
     (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
     (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
     m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
     })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
 
-    ga('create', '{{ ANALYTICS.GOOGLE.GA_SITE_ID }}', 'auto');
+    ga('create', '{{ globals.ANALYTICS.GOOGLE.GA_SITE_ID }}', 'auto');
     ga('set', 'anonymizeIp', true);
     ga('set', 'forceSSL', true);
     ga('send', 'pageview');
